Add digicert gem with an initializer
This CLI is going to utilize the `digicert` gem in underneath. This
commit adds that gem as a dependency with an initializer to set up
the key for the Digicert API.

For now please export `DIGICERT_API_KEY` environment key variables
A better way for this configuration should follow soon.

```sh
bin/digicert hello world
```