Add `digicert` and `terminal-table` gem
Now that we have released our gems to rubygems, so let's start to
use that one and let's also add `terminal-table` gems to display
a nicely formatted output for collections.

This commit also updated the existing `CLI` class and this one
also adds a `Util` class to provide the terminal helper in one
place.