Add `find-orders` with filtering options
This commit adds the interface for the `find-orders` command, now
we can use this one to retrieve the full list of orders or we can
also use `-c` and `-p` options to filter the retrieved orders.

This commit also do some changes on the existing structure as it
was bit different then our current command structure. But we used
most of the functionality we had there just a new structure and
renaming the handler classes. Usages

```ruby
bin/digicert find-orders
bin/digicert find-orders -c "commonname.com", -p "product-name-id"
```