Extract order filtering behavior to a class
We are applying some filters on the retrieved orders from digicert,
currently we are dealing in the `OrderFinder`, but as of our next
feature we will need to implement more filters which is making it
harder to manage from the finder

This commit extract those behavior and added `OrderFilterer`, now
this class will deal with all type of filtering, like we've also
another filter for status now.

This also fetched the latest `digicert` gem from github, which
exposed some interface to use it tests helpers.