Reissue and download a certificate order
Digicer allow us to reissue and download an existing certificate, the
reissuing interface is simple, all we need is create a post request and
digicert will respond with the request details.

But reissuing and instantly trying to download any certificate requires
some additional work. When we submit a reissuing request then digicert
needs some time to actually reissue the order and it doesn't directly
have any interface to fetch recently reissued order / certificate.

This commit adds the interface to perform all of this task in a single
command, if our goal is to reissue an order then we can do that by

```sh
digicert order reissue --order_id 123456789
```
But if we want reissue an order and then fetch and download it then we
can do that by specify the `--output` option, which takes a full path
and it performs the necessary work to fetch the order and then download
it to the provided path.

```sh
digicert order reissue --o 123456789 --output /full/path/to/download
```