Re structure CLI command and sub-command
We are planning to implement more functionality through the `CLI`,
so it would make more sense if we nest the sub command under the
parent command, that will make it easier for user and as well as
easier for us to delegate the behavior to proper handler.

This commit changes the existing structure, For example `order`
related functionality will be nested under the `order` command,
and same will happen will all other interfaces.

In near future we will also implement the `help` interface, that
will provide helpful instructions for each command. Here is an
example of how the new interface will look.

``sh
digicert order list
digicert order find -c "ribosetest.com" -s pending
```