Add interface to retrieve order in interval
When we submit a request to reissue an order then the order might
not be instantly reissued, and it might also takes time to reissue
the order, so we need a way to fetch the updated order in a certain
interval and then maybe download it or do something else.

This commit adds a `OrderRetriever`, which expose `fetch` interface
and that allows us to specify the `number_of_times` and `wait_time`
and then it try to fetch the order as we specify the details.