Generate `CSR` with custom details
This interface adds support to provide the details while creating
a CSR, it adds support for `--san`, `--common_name` along with the
other options, so if we provide the details then it will use those
or else it will use the one from the specified order.

```sh
digicert csr generate --common_name ribosetest.com --order_id 1234 \
 --san test1.ribosetest.com,test2.ribosetest.com, --key key_file
```

Another important note: we also came across an issue for `opensssl`
while we were adding support for `san_names`, which is solved in the
latest version of openssl and that's the reason we are explicitly
specify `openssl` lib and it's version.