Do not show every option in every resource
Currently, we have a `global options` array where we specify all of
the options available, and if we look up the help documentation then
it will provide some miss leading behavior as all options might not
apply in all use case scenario.

This commit improve it a bit, but separating `global` and `local`
options, so if some options are available in all of the resources
then we will add that to `global_options` but if some options are
specific to single resource then we will add that to that specific
resource.

There are still some issue we can improve, like make the options
to `subcommand` specific, but that will require quite some work,
but compare to our use cases I feel it's should be okay for now,
but if we need that then we can always add that later.