Download a specific certificate by its ID
Sometime we might need to download a certificate using it's `ID`. This
commit adds the support to download a certificate by this and it also
implemented the support for `order_id`, so now we can use this whenever
we need to download a certificate by it's id or it's `order_id`.

```sh
digicert certificate download --order_id 654321 --output /downloads
digicert certificate download --certificate_id 12345 --output /downloads
```