Migrate existing option parsing to `thor`
We started pretty small, and as always we tried to keep it simple
without any additional dependencies, but now that we have already
implemented quite some functionalities and we're also planning to
add more in the upcoming days so it's  probably the best time to
migrate our existing CLI to use thor and utilize it's benefits

This commit changes our existing parsing to `thor` and it also
separates interface related behavior to separate commands class
and this should also make it lot easier to manage.

There are quite some part that will require some refactoring or
remove duplications but we will handle those in further PR's.