Expose `digicert` executable in installation
So far we were using the `bin/digicert` directly from the `bin`
directory, but we need to expose this during the installation, so
anyone can use this directly.

To support this we need to add the binary to the user's load path,
and the easiest way to do that is to add it as executable in the
gem specification. This also changes `digicert` gem dependency to
the actual gem instead of GitHub repo.