Delegate shell message printing to thor
Now that we've included `thor`, its probably a good idea to delegate
the shell printing behavior to `thor`, and this also cleanup some of
the existing code that was not using the helper.

One another addition, this commit also adds a support helper which
disabled the message printing from the test suite, so we should not
see any more extra noises in the test log.