Add `rcfile` based Digicert Key configuration
Currently, we are expecting an environment variable as Digicert API
Key, without any easier interface to set it up, So we need a better
solution to handle this and possibly store those in the file system

This commit utilize the `rcfile` based configuration, and provides
an interface to allow user to set their key and then store it to
the home directory as `.digicertrc` and use it when necessary.

This also keeps the environment variable support, so if the user
already has a `DIGICERT_API_KEY` environment variable set up and
has not configured any using `digicert config` then it will use
it and will work the same way.