Add support for remote filtering options
Currently, we are doing the filtering ourself, after pulling all of
the orders we are performing some custom filtering on those, which
is not that efficient and might not be that accurate all the time.

Now that we know digicert has some undocumented support for filter
options so we have implemented the filtering interface and it will
prepare the options properly and pass it two the Digicert API, so
it should provide us much more accurate results now. To use the
filter we can do

```sh
digicert order list --filter 'common_name:ribosetest.com' \
  'status:pending,processing,issued' 'valid_till:<2017-05-11'
```

And most interesting part we don not have manually do anything
anymore, as soon as Digicert adds support for more filters then
we only need to add it to the supported list and it should work

Resolves #32