Fix order reissue/duplicate related issue
The Order `reissue` and `duplication` interface takes an order id
and also the new attributes for the order. The way it supposed to
work is if we add any custom attributes then it will use that with
all others but if we don't then it will use the defaults.

It is working fine when no attributes is provided, but when we add
custom attributes option then it is replacing all of the attributes
with the new one, and which we found from the bug report 46.

This issue was from the digicrt gem, and we already fixed it there
and then this commit upgrades the underlying gem and also add the
required changes for that bug.

Fixes: #46