Remove confusing Key configuration option
In the current version user runs `digicert config API_KEY` to set
the API key, which might seems bit confusing as it does not says
what exactly user is configuring.

This commit changes this configuration options to make it less
confusing and it clearly specify the `api-key` with the config
interface. New way to configure the API key is:

```sh
digicert config api-key YOUR_API_KEY
```

Fixes: #50