Fix certificate download related issue
The `certificate download` command expects us to provide the
order id or certificate id, but does not necessarily force us
to do anything and that's why it was throwing issues like we
have on the Issue: #48.

This commit fixes it by adding a help method that will check
either one of those option are exits, and it not then it will
print out that custom message before calling the API.

This also add another error handling for invalid resource id,
so if the user tries to download an invalid resource then it
will print out the invalid resource related message.