Add `--debug` flag with CLI commands
Currently, in the Digicert Gem we have an option to turn on or off
that `debug_mode`. Based on this one each of the network request
will print out all the important details which is very useful for
debugging.

But, we did not have any support for that in the CLI, so this
commit adds the `--debug` flag with all of the supported commands
so now we can pass that and it will print out the details when
necessary.

Fixes #37