Change CSR's Order Id to be optional
Currently, we are forcing user to provide an order id to generate
new CSR. We are doing it so we can gather the missing informations
using that order id and then use those as a replacement for field
like `common_name` or `organization` details.

But since it's not necessary, so let's change that in a way that
user can choose to provide all the other details by themself.

Fixes: 63