Fix certificate download related issue
For some reason, we missed to include the certificate downloader
and that is what was causing the failures. Our test suite also
seemed to be working unless we ran that specific test file.

Turn's out we were including that downloader file from a spec
file and that's why it was not being reported. This commit fixes
that by adding that downloader to our certificate interface.

Fixes #67